+++ src/DataFixtures/AppFixtures.php
@@ -12,6 +12,6 @@ class AppFixtures extends Fixture
         // $product = new Product();
         // $manager->persist($product);
 
-        // $manager->flush();
+        $manager->flush();
     }
 }
