+++ src/DataFixtures/FacturationFixtures.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace App\DataFixtures;
+
+use Doctrine\Bundle\FixturesBundle\Fixture;
+use Doctrine\Persistence\ObjectManager;
+use App\Entity\Facturation;
+
+class FacturationFixtures extends Fixture
+{
+    public function load(ObjectManager $manager): void
+    {
+        for ($i = 0; $i < 20; $i++) {
+            $facturation = new Facturation();
+            $facturation
+                ->setName("test")
+                ->setStatus("ratio")
+                ->setCreatedAt(new \DateTime())
+                ->setUpdatedAt(new \DateTime())
+            ;
+
+            $manager->persist($facturation);
+        }
+
+        $manager->flush();
+    }
+}
